requirements.txt
Pillow>=10.0.0
pytesseract>=0.3.10
pypdf>=4.0.0
